First Commit (with default main.dart file)
This is the first commit of this project (technically the second one). It has all of the dependencies for the flutter application to work and  the Demo Page of flutter which can be located in main.dart.

Note: Install the Flutter extension in the Visual Studio Code(If you are using Visual Studio Code)  first together with the Flutter SDK to run the application.

It can be ran in any emulator depending on the user. Next step is to create the code for the Baric's Mobile Bar Application.